Remove unneeded duplicate test. Refactor
