--- src/test/java/com/jazeee/ddp/client/AllTests.java
@@ -1,27 +1,0 @@
-/*
- * (c)Copyright 2013-2014 Ken Yee, KEY Enterprise Solutions 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.jazeee.ddp.client;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({ TestDDPBasic.class, TestDDPConnections.class, TestDDPAuth.class, TestDDPCollections.class, TestDDPUsers.class })
-public class AllTests {
-
-}
